server/requirements.txt
Flask==1.0.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
mysqlclient==2.1.1
mysql-connector-python==8.0.31
